Fix: support overwriting routes with regex matcher
Prior to this change, using a regex matcher combined with
`overwriteRoutes: true` would not overwrite the route. This is because
the inspection was comparing regexes with `===` which will always return
`false`. To fix this, object identifiers are coerced to strings before
comparing. Function matchers are not affected by this change.