FIX: Ensure all arguments are passed to base fetch function
This is not a great fix; it ignores that method is only inspected when it's
in the first fetch() parameter, and it glosses over the fact that the
navigator object, for historical reasons, is very constrained in the
responses it gives (which is why the code snagged jsdom in the first
place).

It does, however, solve my immediate need of getting fetch-mock to
interoperate with jsdom.

Fixes wheresrhys/fetch-mock#608